Updating nav and footer

+++ BankBarden/ViewModels/CountryViewModel.cs
@@ -5,6 +5,6 @@
         public int Id { get; set; }
         public string Country { get; set; }
         public int UserCount { get; set; }
-        public int CountryTotalMoney { get; set; }
+        public decimal CountryTotalMoney { get; set; }
     }
 }
